USE fancy slashes for routes
it wasn't the issue.  POSTing was the issue.  Probably just on the SW side.  Whatever, I'm GETting now.  I just wonder if this has privacy considerations like if URLs with params are sent in cleartext at some point vs body parameters which get POSTed.  Will have to ask about that